partial fix for travis config file (no vignette)
This works: dependencies install and R build and check run, however the vignettes do not work and are excluded for now. But at least the tests and other R check portions run.

https://github.com/ShanaScogin/BayesPostEst/issues/45

When trying to build the vignettes, any attempt to fit a Stan or JAGS model fails with some variant of:

Error: C stack usage [XXX] is too close to the limit